add configuration for h2

+++ src/main/java/br/uff/ic/devappcorp/repositories/StudentRepository.java
@@ -3,6 +3,6 @@ package br.uff.ic.devappcorp.repositories;
 import br.uff.ic.devappcorp.entities.Student;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StudentRepository extends CrudRepository<Student, Long> {
+public interface StudentRepository extends CrudRepository<Student, Integer> {
 
 }
